Refactor bookclubs.module folder structure
