+++ src/app/apps/bookclubs/components/bookclub-card/bookclub-card.component.ts
