+++ src/app/apps/bookclubs/components/post-card/post-card.component.ts
