+++ src/app/apps/bookclubs/components/topic-card/topic-card.component.ts
