+++ src/app/apps/bookclubs/pages/topic-details/topic-details.component.ts
